wifi: Init WiFi utils
* Also setup serial
When the board is not able to connect to the targeted WiFi, it will start an AP whose SSID is "ESP32-Door-Setup" and password is "password".
You should connect to the AP manually and go to the web portal whose IP will be printed into the serial output.
In the portal, you can set the SSID and password of the target WiFi and when it is saved, the board will try to connect to it automatically. The connecting result will be printed into the serial.
More infos can be found at tzapu/WiFiManager